The version branch's tip 972a9a6a41 merges --no-ff -S into master 074a12c4ae. It carries the H10
close (STAMP fa18863b94), the post-close batch (61724860b4), H11 (bccf8d977b), H12 with the
announcement (9c9b1a81af; announcement ddc6c03fc9), the README fixes (46696ef0c5) and P4's audit
gate (33b6623eab). OWNER RULING RN-12 / RN-13 (ledger 109208f210): master moves first by a signed
--no-ff merge of the version branch once section 5's parity gates hold, and 1.24.13.1 is then cut
from master, the precedent for every earlier release. Mechanics per the CP2 rulings (ledger
620ba7a2b8) and the CP3 step (ledger 5e19c1844b).

SECTION 5 PARITY GATES, each by its record:
  P1 compile parity    -- discharged at the close STAMP fa18863b94.
  P2 roster parity     -- the section 6 full validated-roster sweep (RN-14; briefed, ledger
                          fb1422104a). THIS MERGE IS PUSHED ONLY AFTER P2 HOLDS.
  P3 behavioral parity -- discharged at the close STAMP fa18863b94.
  P4 hand-own audit    -- the H6 completeness gate PASS, 166 marked paths = 166 rows, class c 0
                          (ledger 15e4a75d9c; re-run at 972a9a6a41, ledger 5e19c1844b).
  P5 ritual rehearsed  -- the isolated-clone rehearsal (RN-7) exercised all five ritual elements,
                          the pre-pack signed tag mint included (ledger 72c59f35ea).

MASTER'S SIDE: the 13 commits the version branch lacked (7 non-merge), every one coordination
tooling or a guard:
  8a218eddd3 / be8e137af5  census: a `converted` mode for tracked converted Go test sources
  20b7457564 / dd18e5e2ab  identifier guards: the owner's PUBLIC handles are an admitted set
  f147fe5973 / facb693040  regen-validation-index: a proof page is backed by a row's own link or
                           an exclusion class (blob-equal on both sides; the version branch
                           materialised it at f545b18d4d, so it merges without a change)
  9210bc85f4 / 998fc49d03  census: Go's hex byte escape admitted (reverted by 17aff5dc5d)
  17aff5dc5d, 3d042db083 / 20eb0af70e  census: converted mode mirrors the upstream-fixture predicate
  33f5f11297 / 074a12c4ae  census: converted mode admits a //go:embed payload its csproj names
Files: .claude/coord-scripts/coord-identifier-{census.sh,hashes.txt,patterns.txt},
.claude/rules/docs-records.md, docs/phase4/hopA-inputs/regen-validation-index.py and
src/go2cs/internal/repoguard/fleetIdentifierCensus_test.go. The version side touched none of them
except the last two since the merge base 9e12c3e7d0.

ONE CONFLICT, resolved as the UNION the ruling names:
src/go2cs/internal/repoguard/fleetIdentifierCensus_test.go, two hunks. Master's public-handle admit
(20b7457564: an `admit` index threaded through scanFleetIdentifiers, scanFleetTree and
fleetLineHasDeniedToken) meets the version tip's //go:embed payload admit (de436207c9:
scanFleetIdentifiers becomes a wrapper over scanFleetContent, whose caller decides `upstream`).
  hunk 1: scanFleetIdentifiers keeps master's signature (denied, admit) and wraps
          scanFleetContent(path, content, denied, admit, fleetIsUpstreamFixture(path));
          scanFleetContent takes both (denied, admit ... , upstream bool).
  hunk 2: scanFleetTree calls scanFleetContent(p, content, denied, admit,
          fleetIsUpstreamFixture(p) || payloads[p]).
  outside the markers: the version tip's new TestFleetIdentifierEmbedPayloadsAreAdmittedAsFixtures
          called scanFleetTree(dir, rel, denied) at the old arity, which the union does not
          compile; it passes `nil` for admit, as master's own controls do.
Both suites keep their own tests: TestFleetPublicHandleAdmitIsBounded and
TestFleetIdentifierHashFileMatchesTheGoLists (master), TestFleetIdentifierEmbedPayloadsAreAdmittedAsFixtures
and TestFleetEmbedPayloadDerivationReadsTheTree (version). Every other path merged without a
conflict: the corpus, docs/README.md, docs/ValidatedTestPackages.md and docs/NEWS.md come from the
version side (master did not touch them since the merge base); CLAUDE.md and .claude/rules were each
edited on one side only.

Prepared on the i7 in a throwaway worktree; the gates at this merge are reported to COORD.

+++ src/core/math/big/decimal.cs
@@ -89,7 +89,7 @@ internal static void init(this ж<@decimal> Ꮡx, nat m, nint shift) {
     x.mant = appendꓸꓸꓸ(x.mant[..0], s[..(int)(n)]);
     // Do any (remaining) shift right in decimal representation.
     if (shift < 0) {
-        while (shift < -maxShift) {
+        while (shift < (nint)(-maxShift)) {
             shr(ref (Ꮡx).DerefOrNull(), maxShift);
             shift += maxShift;
         }
